Struct proc_macro_warning::DeprecatedWarningBuilder

pub struct DeprecatedWarningBuilder { /* private fields */ }

Gradually build a deprecation Warning.

Example

use proc_macro_warning::Warning;

let warning = Warning::new_deprecated("my_macro")
    .old("my_macro()")
    .new("my_macro::new()")
    .help_link("https:://example.com")
    // Normally you use the input span, but this is an example:
    .span(proc_macro2::Span::call_site())
    .build();

let mut warnings = vec![warning];
// When adding more, you will need to build each with `.index`.

// In a proc macro you can expand them in a private module:
quote::quote! {
    mod warnings {
        #(
            #warnings
        )*
    }
};

Implementations

impl DeprecatedWarningBuilder

pub fn from_title(title: &str) -> DeprecatedWarningBuilder

Create a new deprecated warning builder with the given title.

The title must be unique for each warning.

pub fn index(self, index: usize) -> DeprecatedWarningBuilder

Set an optional index in case that a warning appears multiple times.

Must be set if a warning appears multiple times.

pub fn old(self, old: &str) -> DeprecatedWarningBuilder

The old deprecated way of doing something.

Should complete the sentence “It is deprecated to …”.

pub fn new(self, new: &str) -> DeprecatedWarningBuilder

The new way of doing something.

Should complete the sentence “Please instead …”.

pub fn help_link(self, link: &str) -> DeprecatedWarningBuilder

A help link for the user to explain the transition and justification.

pub fn help_links(self, links: &[&str]) -> DeprecatedWarningBuilder

Multiple help links for the user to explain the transition and justification.

pub fn span(self, span: Span) -> DeprecatedWarningBuilder

The span of the warning.

pub fn maybe_build(self) -> Result<Warning, String>

Build the warning.

pub fn build(self) -> Warning

Unwraps Self::maybe_build for convenience.